+++ templates/template2/run.sh
@@ -16,4 +16,8 @@ sleep 2
 python3 ./templates/template2/build_videos.py $folder_path
 python3 ./templates/template2/add_transcription.py $folder_path
 
-kill $XVFB_PID
+if ps -p $XVFB_PID > /dev/null
+
+then
+    kill $XVFB_PID
+fi
